@@ -25,19 +25,14 @@
 		}
 	},
     "require": {
-		"php": "^7.2 || ^8.0",
-		"guzzlehttp/guzzle": "^6.0 || ^7.0",
-		"typo3/cms-core": "^10.0 || ^11.0",
-		"doctrine/dbal": "~2.13.1"
+		"php": "^8.1",
+		"typo3/cms-backend": "^11.5 || ^12.4",
+		"typo3/cms-frontend": "^11.5 || ^12.4"
 	},
 	"require-dev": {
-		"phpstan/phpstan": "^0.12.11",
-		"phpunit/phpunit": "^8.4",
-		"typo3/cms-core": "^10",
-		"typo3/cms-frontend": "^10",
-		"typo3/coding-standards": "^0.5.5",
-		"typo3/tailor": "^1.0",
-		"typo3/testing-framework": "^6"
+		"saschaegerer/phpstan-typo3": "^1.8",
+		"typo3/coding-standards": "^0.7",
+		"typo3/tailor": "^1.0"
 	},
 	"extra": {
 		"typo3/cms": {
@@ -46,14 +41,6 @@
 			"app-dir": ".Build",
 			"web-dir": ".Build/Web"
 		}
-	},
-	"scripts": {
-		"post-autoload-dump": [
-			"TYPO3\\TestingFramework\\Composer\\ExtensionTestEnvironment::prepare",
-			"if [ ! -e .Build/Web/vendor ]; then cd .Build/Web  && ln -s ../vendor && cd -; fi",
-			"if [ ! -e .Build/Web/typo3conf/ext/json_response ]; then cd .Build/Web/typo3conf/ext && ln -s ../../../vendor/typo3/testing-framework/Resources/Core/Functional/Extensions/json_response && cd -; fi",
-			"if [ ! -e .Build/Web/typo3conf/sites ]; then cd .Build/Web/typo3conf && ln -s ../../../Build/sites && cd -; fi"
-		]
 	},
 	"autoload": {
 		"psr-4": {
