@@ -4,7 +4,7 @@ $EM_CONF[$_EXTKEY] = [
     'title' => 'Reverse Proxy - Cache Manager',
     'description' => 'A flexible and generic way to track the pages that are cached by a reverse proxy like nginx HA or a CDN.',
     'category' => 'fe',
-    'version' => '3.4.2',
+    'version' => '4.0.0',
     'state' => 'stable',
     'clearcacheonload' => 1,
     'author' => 'Benjamin Mack',
@@ -12,7 +12,7 @@ $EM_CONF[$_EXTKEY] = [
     'author_company' => 'b13 GmbH',
     'constraints' => [
         'depends' => [
-            'typo3' => '10.4.0-11.5.99',
+            'typo3' => '11.5.0-12.4.99',
         ],
         'conflicts' => [
         ],
